+++ tcvectordb/model/collection.py
@@ -2,7 +2,7 @@ from typing import Dict, List, Optional, Any, Union
 
 from numpy import ndarray
 
-from tcvectordb.model.index import IndexField, VectorIndex, FilterIndex, SparseVector
+from tcvectordb.model.index import IndexField, VectorIndex, FilterIndex, SparseVector, SparseIndex
 
 from tcvectordb import exceptions
 from .document import Document, Filter, AnnSearch, KeywordSearch, Rerank
@@ -1055,13 +1055,13 @@ class Collection():
         return res.data()
 
     def modify_vector_index(self,
-                            vector_indexes: List[VectorIndex],
+                            vector_indexes: List[Union[VectorIndex, SparseIndex]],
                             rebuild_rules: Optional[dict] = None,
                             timeout: Optional[float] = None) -> dict:
         """Adjust vector index parameters.
 
         Args:
-            vector_indexes (List[FilterIndex]): The vector fields to adjust
+            vector_indexes (List[Union[VectorIndex, SparseIndex]]): The vector fields to adjust
             rebuild_rules (dict): Specified rebuild rules.
                     This interface will trigger a rebuild after adjusting the parameters.
                     For example: {"drop_before_rebuild": True , "throttle": 1}