+++ tcvectordb/rpc/client/stub.py
@@ -900,7 +900,7 @@ class RPCVectorDBClient(VectorDBClient):
     def modify_vector_index(self,
                             database_name: str,
                             collection_name: str,
-                            vector_indexes: List[VectorIndex],
+                            vector_indexes: List[Union[VectorIndex, SparseIndex]],
                             rebuild_rules: Optional[dict] = None,
                             timeout: Optional[float] = None) -> dict:
         """Adjust vector index parameters.
@@ -908,7 +908,7 @@ class RPCVectorDBClient(VectorDBClient):
         Args:
             database_name (str): The name of the database where the collection resides.
             collection_name (str): The name of the collection
-            vector_indexes (List[FilterIndex]): The vector fields to adjust
+            vector_indexes (List[Union[VectorIndex, SparseIndex]]): The vector fields to adjust
             rebuild_rules (dict): Specified rebuild rules.
                     This interface will trigger a rebuild after adjusting the parameters.
                     For example: {"drop_before_rebuild": True , "throttle": 1}